Conflicts:
	palladian-core/src/main/java/ws/palladian/classification/utils/ClassificationUtils.java
	palladian-core/src/test/java/ws/palladian/classification/universal/UniversalClassifierTest.java

Former-commit-id: c8dcd0a28eac704f0c1b320248ff100bfaf5fe43

+++ palladian-retrieval/src/main/java/ws/palladian/retrieval/search/BaseBingSearcher.java
@@ -227,7 +227,7 @@ public abstract class BaseBingSearcher<R extends WebResult> extends WebSearcher<
     }
 
     @Override
-    public int getTotalResultCount(String query, Language language) throws SearcherException {
+    public long getTotalResultCount(String query, Language language) throws SearcherException {
         throw new SearcherException("Getting the total result count is not supported in the new Bing API.");
     }
